requirements-secure.txt
requests>=2.28.0
Flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
polygon-api-client>=1.7.0
urllib3>=1.26.0
psutil>=5.8.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
lxml>=5.2.0
pydantic>=2.8.0
SQLAlchemy>=2.0.0
apscheduler>=3.10.0
click>=8.1.0
playwright>=1.48.0
